requirements.txt
visbrain
pandas
dipy
numpy==1.26.4 
nibabel
matplotlib
pyyaml
scikit-learn
prettytable
nilearn
nipype
scikit-image
spektral
optuna